requirements.txt
# requirements.txt
pyautogui>=0.9.54
pandas>=2.3.3
pyperclip>=1.8.2
openpyxl>=3.1.2
clipboard>=0.0.4
Pillow>=10.0.0
keyboard>=0.13.5
requests>=2.32.5
quart >= 0.20.0
dotenv>=0.9.9
flet-webview>=0.1.0
flet[all]